🔖 chore(package.json): bump version from 2.0.3 to 2.1.0 ✨ feat(package.json): add 'iphone-dynamic' and 'iphone-dynamic-island' to keywords ➕ feat(package.json): add 'react-native-device-info' to peerDependencies
The version bump is due to the addition of new features. The new keywords 'iphone-dynamic' and 'iphone-dynamic-island' have been added to improve the discoverability of the package. The 'react-native-device-info' package has been added to peerDependencies to ensure that it is installed alongside this package, as it is now a necessary dependency for the library to function correctly.